+++ src/os_cmd/presets/cargo_build/cargo_profile.rs
@@ -2,12 +2,20 @@ use crate::os_cmd::MiniStr;
 
 #[derive(Debug, Clone, PartialEq, Eq)]
 /// cargo profile name: e.g., build, dev
+#[derive(Default)]
 pub enum CargoProfile {
   /// "release"
+  #[default]
   Release,
   /// "dev"
   Debug,
   Custom(MiniStr),
+}
+
+impl core::fmt::Display for CargoProfile {
+  fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+    f.write_str(self.as_str())
+  }
 }
 
 impl From<&str> for CargoProfile {
@@ -33,12 +41,6 @@ impl CargoProfile {
 impl AsRef<str> for CargoProfile {
   fn as_ref(&self) -> &str {
     self.as_str()
-  }
-}
-
-impl Default for CargoProfile {
-  fn default() -> Self {
-    Self::Release
   }
 }